/categoria

+++ src/index.ts
@@ -10,6 +10,27 @@ import prisma from '@utils/prisma'
 import { Income } from '@prisma/client'
 dotenv.config()
 
+type IncomeJSON = {
+  amount: number
+  currency: 'HNL' | 'USD'
+  source: string
+} | {
+  error: string
+  reset?: boolean
+}
+
+type CategoryJSON = {
+  description: string
+  budget: number
+  currency: 'HNL' | 'USD'
+  isFixed: boolean
+  dueDate?: number
+  notes: string
+} | {
+  error: string
+  reset?: boolean
+}
+
 if (process.env.botToken === undefined) {
   throw new Error('botToken is not defined')
 }
@@ -22,17 +43,6 @@ if (process.env.timezone === undefined) {
   throw new Error('timezone is not defined')
 }
 
-if (process.env.DOLLAR_TO_HNL === undefined) {
-  throw new Error('DOLLAR_TO_HNL is not defined')
-}
-
-if (process.env.HNL_TO_DOLLAR === undefined) {
-  throw new Error('HNL_TO_DOLLAR is not defined')
-}
-
-const dollarToHNL = parseFloat(process.env.DOLLAR_TO_HNL)
-const hnlToDollar = parseFloat(process.env.HNL_TO_DOLLAR)
-
 dayjs.extend(utc)
 dayjs.extend(timezone)
 
@@ -43,7 +53,6 @@ const token = process.env.botToken
 const bot = new TelegramBot(token, { polling: true })
 
 bot.on('message', async (msg) => {
-  const today = dayjs().tz(process.env.timezone).format('YYYY-MM-DD HH:mm:ss')
   if (!userWhitelist.includes(msg.chat.id)) {
     bot.sendMessage(msg.chat.id, `You, ${msg.chat.id} are not authorized to use this bot.`)
     return
@@ -54,8 +63,13 @@ bot.on('message', async (msg) => {
     return
   }
 
+  bot.sendChatAction(msg.chat.id, 'typing')
+  const today = dayjs().tz(process.env.timezone).format('YYYY-MM-DD HH:mm:ss')
+  const conversion = await prisma.conversion.findFirst()
+  const dollarToHNL = conversion?.dollarToHNL || 24.6
+  const hnlToDollar = conversion?.hnlToDollar || 0.04
+
   if (msg.text === '/start') {
-    bot.sendChatAction(msg.chat.id, 'typing')
     await bot.setMyCommands([{
       command: 'estado',
       description: 'Crear o edita tus estados de cuentas mensuales.'
@@ -65,6 +79,9 @@ bot.on('message', async (msg) => {
     }, {
       command: 'categoria',
       description: 'Crear o edita tus categorías de gastos.'
+    }, {
+      command: 'resumen',
+      description: 'Ver un resumen de tus estado de cuenta actual.'
     }], {
       scope: {
         type: msg.chat.type === 'group' ? 'all_group_chats' : 'chat',
@@ -88,6 +105,7 @@ bot.on('message', async (msg) => {
 
     await prisma.user.create({
       data: {
+        fullName: msg.chat.first_name || msg.chat.username || '',
         chatId: msg.chat.id,
         chatSubject: '',
         chatSubSubject: '',
@@ -134,25 +152,7 @@ bot.on('message', async (msg) => {
     }
   }) || []
 
-  if (userText === '/No') {
-    bot.sendChatAction(msg.chat.id, 'typing')
-    // Reset chat
-    await prisma.user.update({
-      where: {
-        chatId: msg.chat.id
-      },
-      data: {
-        chatHistory: [],
-        chatSubject: '',
-        chatSubSubject: ''
-      }
-    })
-    await bot.sendMessage(msg.chat.id, 'Entendido. ¿En qué más puedo ayudarte?')
-  }
-
   if (userText === '/ingreso') {
-    bot.sendChatAction(msg.chat.id, 'typing')
-
     if (!user.statementIdSet) {
       await bot.sendMessage(msg.chat.id, 'Primero debes crear un estado de cuenta.\nUsa /estado para crear uno.')
       return
@@ -173,17 +173,35 @@ bot.on('message', async (msg) => {
   }
 
   if (user.chatSubject === 'income') {
-    bot.sendChatAction(msg.chat.id, 'typing')
-
     if (user.chatSubSubject === 'view') {
       if (userText.startsWith('/')) {
         try {
           const index = parseInt(userText.replace('/', '')) - 1
-          const incomes = await prisma.income.findMany()
+          const incomeSelected = await prisma.income.findFirst({
+            skip: index,
+            where: {
+              statementId: user.statementIdSet as string
+            }
+          })
+
+          if (!incomeSelected) {
+            await bot.sendMessage(msg.chat.id, 'No se encontró el ingreso.')
+            await prisma.user.update({
+              where: {
+                chatId: msg.chat.id
+              },
+              data: {
+                chatHistory: [],
+                chatSubject: '',
+                chatSubSubject: ''
+              }
+            })
+            return
+          }
 
           await prisma.income.delete({
             where: {
-              id: incomes[index].id
+              id: incomeSelected.id
             }
           })
           await bot.sendMessage(msg.chat.id, 'Ingreso eliminado correctamente.')
@@ -191,7 +209,7 @@ bot.on('message', async (msg) => {
           await bot.sendMessage(msg.chat.id, 'No se pudo eliminar el ingreso. Inténtalo de nuevo.')
         }
 
-        prisma.user.update({
+        await prisma.user.update({
           where: {
             chatId: msg.chat.id
           },
@@ -229,7 +247,12 @@ bot.on('message', async (msg) => {
             chatHistory: []
           }
         })
-        const incomes = await prisma.income.findMany()
+
+        const incomes = await prisma.income.findMany({
+          where: {
+            statementId: user.statementIdSet as string
+          }
+        })
 
         if (incomes.length === 0) {
           await bot.sendMessage(msg.chat.id, 'No tienes ingresos registrados.')
@@ -260,37 +283,6 @@ bot.on('message', async (msg) => {
     }
 
     if (user.chatSubSubject === 'create') {
-      if (userText === '/Si') {
-        const newData = JSON.parse(chatHistory[chatHistory.length - 1].content)
-
-        await prisma.income.create({
-          data: {
-            // @ts-ignore
-            statementId: user.statementIdSet,
-            // @ts-ignore
-            amount: parseFloat(newData.amount),
-            // @ts-ignore
-            currency: newData.currency,
-            // @ts-ignore
-            source: newData.source,
-          }
-        })
-
-        await prisma.user.update({
-          where: {
-            chatId: msg.chat.id
-          },
-          data: {
-            chatHistory: [],
-            chatSubject: '',
-            chatSubSubject: ''
-          }
-        })
-
-        await bot.sendMessage(msg.chat.id, 'Ingreso creado correctamente.')
-        return
-      }
-
       const botAI = await openAi.chat.completions.create({
         messages: [{
           role: 'system',
@@ -318,7 +310,7 @@ bot.on('message', async (msg) => {
       })
 
       const botMessage = botAI.choices[0].message.content?.trim() || '{"error": "No se pudo procesar la información.", "reset": true}'
-      const botMessageJSON = JSON.parse(botMessage)
+      const botMessageJSON: IncomeJSON = JSON.parse(botMessage)
 
       if ('reset' in botMessageJSON) {
         await prisma.user.update({
@@ -326,6 +318,8 @@ bot.on('message', async (msg) => {
             chatId: msg.chat.id
           },
           data: {
+            chatSubject: '',
+            chatSubSubject: '',
             chatHistory: []
           }
         })
@@ -352,26 +346,31 @@ bot.on('message', async (msg) => {
         return
       }
 
-      await bot.sendMessage(msg.chat.id, botMessage)
       await prisma.user.update({
         where: {
           chatId: msg.chat.id
         },
         data: {
-          chatHistory: {
-            push: [
-              JSON.stringify({ role: 'user', content: userText }),
-              JSON.stringify({ role: 'assistant', content: botMessage })
-            ]
-          }
+          chatSubject: '',
+          chatSubSubject: '',
+          chatHistory: []
         }
       })
-      await bot.sendMessage(msg.chat.id, "¿Estás de acuerdo con estos datos?\n/Si\n/No")
+
+      const newIncome = await prisma.income.create({
+        data: {
+          statementId: user.statementIdSet as string,
+          amount: botMessageJSON.amount,
+          currency: botMessageJSON.currency,
+          source: botMessageJSON.source,
+        }
+      })
+
+      await bot.sendMessage(msg.chat.id, `Ingreso ${newIncome.source}, creado correctamente.`)
     }
   }
 
   if (userText === '/estado') {
-    bot.sendChatAction(msg.chat.id, 'typing')
     await prisma.user.update({
       where: {
         chatId: msg.chat.id
@@ -437,6 +436,8 @@ bot.on('message', async (msg) => {
             chatId: msg.chat.id
           },
           data: {
+            chatSubject: '',
+            chatSubSubject: '',
             chatHistory: []
           }
         })
@@ -502,9 +503,9 @@ bot.on('message', async (msg) => {
                   createMany: {
                     data: prevStatement.categories.map(c => ({
                       currency: c.currency,
-                      debt: c.debt,
                       description: c.description,
-                      limit: c.limit,
+                      budget: c.budget,
+                      isFixed: c.isFixed,
                       dueDate: c.dueDate,
                       notes: c.notes
                     }))
@@ -536,7 +537,11 @@ bot.on('message', async (msg) => {
           }
         })
 
+        const monthInSpanish = dayjs().locale('es').month(newStatement?.month - 1).format('MMMM')
+        const year = newStatement?.year
+
         await bot.sendMessage(msg.chat.id, 'Estado de cuenta creado correctamente.')
+        await bot.sendMessage(msg.chat.id, `Estado de cuenta actual: ${monthInSpanish} ${year}`)
         return
       }
 
@@ -552,18 +557,274 @@ bot.on('message', async (msg) => {
         }
       })
 
+      const monthInSpanish = dayjs().locale('es').month(statement?.month - 1).format('MMMM')
+      const year = statement?.year
       await bot.sendMessage(msg.chat.id, 'Estado de cuenta encontrado.')
-
-      await bot.sendMessage(msg.chat.id, botMessage)
+      await bot.sendMessage(msg.chat.id, `Estado de cuenta actual: ${monthInSpanish} ${year}`)
     }
   }
 
   if (userText === '/categoria') {
-    bot.sendChatAction(msg.chat.id, 'typing')
-
     if (!user.statementIdSet) {
       await bot.sendMessage(msg.chat.id, 'Primero debes crear un estado de cuenta.\nUsa /estado para crear uno.')
       return
+    }
+
+    await prisma.user.update({
+      where: {
+        chatId: msg.chat.id
+      },
+      data: {
+        chatSubject: 'category',
+        chatSubSubject: 'ask',
+        chatHistory: []
+      }
+    })
+    await bot.sendMessage(msg.chat.id, '¿Que deseas hacer?\n/crear una categoria\n/ver tus categorias')
+    return
+  }
+
+  if (user.chatSubject === 'category') {
+    if (userText === '/crear') {
+      const botReply = '¿Cuál es la descripción de la categoría?'
+      await prisma.user.update({
+        where: {
+          chatId: msg.chat.id
+        },
+        data: {
+          chatSubSubject: 'create-description',
+          chatHistory: {
+            push: [JSON.stringify({ role: 'assistant', content: botReply })]
+          }
+        }
+      })
+      await bot.sendMessage(msg.chat.id, botReply)
+      return
+    }
+
+    if (userText === '/ver') {
+      await prisma.user.update({
+        where: {
+          chatId: msg.chat.id
+        },
+        data: {
+          chatHistory: [],
+          chatSubSubject: 'view'
+        }
+      })
+
+      const categories = await prisma.category.findMany({
+        where: {
+          statementId: user.statementIdSet as string
+        }
+      })
+
+      if (categories.length === 0) {
+        await bot.sendMessage(msg.chat.id, 'No tienes categorías registradas.')
+        return
+      }
+
+      const categoriesText = categories.map((category, i) => {
+        return `/${i + 1}. <b>${category.description}</b>\n${category.budget} ${category.currency}`
+      }).join('\n\n')
+
+      await bot.sendMessage(msg.chat.id, `Presiona el /# para editar o eliminar.\n\n${categoriesText}`, { parse_mode: 'HTML' })
+    }
+
+    if (user.chatSubSubject === 'view') {
+      if (userText.startsWith('/')) {
+        try {
+          const index = parseInt(userText.replace('/', '')) - 1
+          const category = await prisma.category.findFirst({
+            skip: index,
+            where: {
+              statementId: user.statementIdSet as string
+            }
+          })
+
+          if (!category) {
+            await bot.sendMessage(msg.chat.id, 'No se encontró la categoría.')
+            await prisma.user.update({
+              where: {
+                chatId: msg.chat.id
+              },
+              data: {
+                chatHistory: [],
+                chatSubject: '',
+                chatSubSubject: ''
+              }
+            })
+            return
+          }
+
+          await prisma.user.update({
+            where: {
+              chatId: msg.chat.id
+            },
+            data: {
+              chatHistory: [],
+              chatSubSubject: 'view-ask'
+            }
+          })
+          await bot.sendMessage(msg.chat.id, `<b>${category.description}</b>\n${numeral(category.budget).format('0,0.00')} ${category.currency}\nFija: ${category.isFixed ? `Si\nCada ${category.dueDate}` : 'No'}${category.notes ? `\n<blockquote>${category.notes}</blockquote>` : ''}\n\n/eliminar\n/editar`, { parse_mode: 'HTML' })
+        } catch (error) {
+          await prisma.user.update({
+            where: {
+              chatId: msg.chat.id
+            },
+            data: {
+              chatHistory: [],
+              chatSubject: '',
+              chatSubSubject: ''
+            }
+          })
+          await bot.sendMessage(msg.chat.id, 'Ocurrió un error. Inténtalo de nuevo.')
+        }
+      }
+    }
+
+    if (user.chatSubSubject === 'create-description') {
+      const botReply = '¿Cuál es el presupuesto y su moneda de la categoría?'
+      await prisma.user.update({
+        where: {
+          chatId: msg.chat.id
+        },
+        data: {
+          chatSubSubject: 'create-budget',
+          chatHistory: {
+            push: [
+              JSON.stringify({ role: 'user', content: userText }),
+              JSON.stringify({ role: 'assistant', content: botReply }),
+            ]
+          }
+        }
+      })
+
+      await bot.sendMessage(msg.chat.id, botReply)
+      return
+    }
+
+    if (user.chatSubSubject === 'create-budget') {
+      const botReply = '¿Es un gasto fijo? Si lo es, ¿cuál es la fecha de vencimiento?'
+      await prisma.user.update({
+        where: {
+          chatId: msg.chat.id
+        },
+        data: {
+          chatSubSubject: 'create-isFixed',
+          chatHistory: {
+            push: [
+              JSON.stringify({ role: 'user', content: userText }),
+              JSON.stringify({ role: 'assistant', content: botReply }),
+            ]
+          }
+        }
+      })
+
+      await bot.sendMessage(msg.chat.id, botReply)
+      return
+    }
+
+    if (user.chatSubSubject === 'create-isFixed') {
+      const botReply = 'Agrega alguna nota adicional.'
+      await prisma.user.update({
+        where: {
+          chatId: msg.chat.id
+        },
+        data: {
+          chatSubSubject: 'create-end',
+          chatHistory: {
+            push: [
+              JSON.stringify({ role: 'user', content: userText }),
+              JSON.stringify({ role: 'assistant', content: botReply }),
+            ]
+          }
+        }
+      })
+
+      await bot.sendMessage(msg.chat.id, botReply)
+      return
+    }
+
+    if (user.chatSubSubject === 'create-end') {
+      const botAI = await openAi.chat.completions.create({
+        messages: [{
+          role: 'system',
+          content: `Today is: ${today}`
+        }, {
+          role: 'system',
+          content: 'Reply in spanish'
+        }, {
+          role: 'system',
+          content: `Your job is to get category description, budget amount, currency, isFixed, dueDate and notes from the user.`
+        }, {
+          role: 'system',
+          content: `The currency can be HNL (the user can type L, Lempiras or HNL) or USD (the user can type as $ or Dollars).`
+        }, {
+          role: 'system',
+          content: 'You will return these data in JSON format: { "description": <description>, "budget": <amount>, "currency": "HNL" or "USD", "isFixed": <true | false>, "dueDate": <if isFixed is true, from 0 to 31>, "notes": <notes or empty string> } or { "error": "error message" }'
+        }, ...chatHistory, {
+          role: 'user',
+          content: userText
+        }],
+        model: 'gpt-4-1106-preview',
+        response_format: { type: 'json_object' },
+        temperature: 0.2,
+        max_tokens: 300
+      })
+
+      const botMessage = botAI.choices[0].message.content?.trim() || '{"error": "No se pudo procesar la información.", "reset": true}'
+      const botMessageJSON: CategoryJSON = JSON.parse(botMessage)
+
+      if ('reset' in botMessageJSON) {
+        await prisma.user.update({
+          where: {
+            chatId: msg.chat.id
+          },
+          data: {
+            chatSubject: '',
+            chatSubSubject: '',
+            chatHistory: []
+          }
+        })
+
+        await bot.sendMessage(msg.chat.id, 'No se pudo procesar la información. Inténtalo de nuevo.')
+        return
+      }
+
+      if ('error' in botMessageJSON) {
+        await bot.sendMessage(msg.chat.id, `${botMessageJSON.error}\nInténtalo de nuevo con /categoria.`)
+        await prisma.user.update({
+          where: {
+            chatId: msg.chat.id
+          },
+          data: {
+            chatSubject: '',
+            chatSubSubject: '',
+            chatHistory: []
+          }
+        })
+        return
+      }
+
+      try {
+        const newCategory = await prisma.category.create({
+          data: {
+            description: botMessageJSON.description,
+            budget: botMessageJSON.budget,
+            currency: botMessageJSON.currency,
+            isFixed: botMessageJSON.isFixed,
+            dueDate: botMessageJSON.dueDate,
+            notes: botMessageJSON.notes || '',
+            statementId: user.statementIdSet as string
+          }
+        })
+        await bot.sendMessage(msg.chat.id, `Categoría ${newCategory.description}, creada correctamente.`)
+        return
+      } catch (error) {
+        await bot.sendMessage(msg.chat.id, 'No se pudo crear la categoría. Inténtalo de nuevo.')
+        return
+      }
     }
   }
 })
